python-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
openpyxl==3.1.2
asyncpg==0.29.0
reportlab==4.0.4
psutil==5.9.0
